Added conversion functions between arrays with and without boundary (#49)
* Added conversion functions between arrays with and without boundary

Those functions use views and have a negligable overhead.

Also updated the right hand side function to minimize copying

* Fix duplicated lines
